6-sizwe

+++ 0x00-hello_world/6-size.c
@@ -5,12 +5,11 @@
  */
 int main(void)
 {
-	printf("Size of a char: %u byte(s)", (unsigned int)sizeof(char));
-	printf("Size of an int: %u byte(s)", (unsigned int)sizeof(int));
-	printf("Size of a long int: %u byte(s)", (unsigned int)sizeof(long int));
-	printf("Size of a long long int: %u byte(s)", (unsigned int)
-	sizeof(long long int));
-	printf("Size of a float: %lu byte(s)", (unsigned int)sizeof(float));
+	printf("Size of a char: %lu byte(s)", sizeof(char));
+	printf("Size of an int: %lu byte(s)", sizeof(int));
+	printf("Size of a long int: %lu byte(s)", sizeof(long int));
+	printf("Size of a long long int: %lu byte(s)", sizeof(long long int));
+	printf("Size of a float: %lu byte(s)", sizeof(float));
 	return (0);
 
 }
